+++ macros/spells/moonbeam.js
@@ -76,7 +76,7 @@ if (args[0] === "on") {
     const spell = {
       name: castItemName,
       type: "spell",
-      data: {
+      system: {
         description: DAEItem.system.description,
         activation: { type: "action", },
         ability: DAEItem.system.ability,
@@ -91,6 +91,7 @@ if (args[0] === "on") {
         scaling: { mode: "none", formula: "", },
       },
       img: DAEItem.img,
+      flags: { ddbimporter: { ignoreItemUpdate: true } },
       effects: [],
     };
     await targetActor.createEmbeddedDocuments("Item", [spell]);